requirements.txt
baidu-aip==2.2.13.0
certifi==2019.3.9
chardet==3.0.4
Faker==1.0.6
future==0.17.1
idna==2.8
itchat==1.2.32
lxml==4.3.3
Pillow==6.0.0
pymongo==3.8.0
pypng==0.0.19
PyQRCode==1.2.1
python-dateutil==2.8.0
PyYAML==5.1
requests==2.21.0
six==1.12.0
text-unidecode==1.2
urllib3==1.24.3
wxpy==0.3.9.8
